refactor: pages 폴더 중 LoginPage 폴더에 있던 것을 components 폴더로 이동
